Revert "Qlora 8bit"

+++ glm4_lora_evalute.py
@@ -30,10 +30,8 @@ if __name__ == '__main__':
     tokenizer = AutoTokenizer.from_pretrained(mode_path, trust_remote_code=True)
 
     # 加载模型
-    model = AutoModelForCausalLM.from_pretrained(mode_path, device_map="auto", torch_dtype=torch.half,
+    model = AutoModelForCausalLM.from_pretrained(mode_path, device_map="auto", torch_dtype=torch.bfloat16,
                                                  trust_remote_code=True).eval()
-
-    # model = model.half().cuda()
 
     # 加载lora权重
     model = PeftModel.from_pretrained(model, model_id=lora_path)
